Amount changed from String to Int in Grocery data class

+++ app/src/main/java/com/example/myapplicationasdsds/MainActivity.kt
@@ -54,7 +54,7 @@ class MainActivity : ComponentActivity() {
                             //.background(color = Color.LightGray)
                         ){
                             name, amount ->
-                            viewModel.addGrocery(name, amount)
+                            viewModel.addGrocery(name, amount.toInt())
                         }
                 }
             }
